Rename to draggable.js to dnd.js
